fix: Remove input listener from card SDK
This prevents a bug causing the need to click the transaction
button twice. Now the user input values are requested from
the view when a transaction is initiated rather when an
input field value changes. Devs that want to push values
before a transaction is executed can do so using
tokenscript.action.setProps provided by the card SDK.