feat: Add card SDK method for named TX execution
This commit extends the existing tokenscript.action.executeTransaction
method to allow the execution of named transactions defined in the top
level of the tokenscript XML. There is also a few fixes to allow the correct
reloading of token context data after the transaction is executed, without
reloading the entire card.

This allows cards to become an entry point that can execute many declarative
transactions, rather than cards being tied to a specific TX. There is more
work needed to allow awaiting transactions and adding official integration
of transactions events into the engine.