feat: Allow complex input types for ethereum:transaction and ethereum:call args
This commit allows complex input types for function inputs, such as arrays or structs.
When ts:struct is used as an argument input, it's local-ref is used to get
the correct input type from the ABI supplied within the contract definition.

Complex arguments can then be set using setProps() within the JS card.